% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add variable labels for data set.}
\usage{
add_labels(
  x,
  vars = c("sex", "cancer", "areacode"),
  label_type = "full",
  lang = "zh",
  as_factor = TRUE
)
}
\arguments{
\item{x}{A data frame or tibble to be labeled.}

\item{vars}{A character vector of variable names to label. Default includes
\code{"sex"}, \code{"cancer"}, and \code{"areacode"}.}

\item{label_type}{Type of the label used ("full" or "abbr").}

\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}

\item{as_factor}{Logical, indicate whether output value as factor.}
}
\value{
A data frame or tibble with labeled variables and reordered columns.
}
\description{
\code{add_labels()} adds labels for selected variables such as \code{sex}, \code{cancer},
and \code{areacode} in the dataset based on the specified language and label type.
}
\examples{
data("canregs")
asr <- create_asr(canregs[[1]], year, sex, cancer)
asr <- add_labels(asr, label_type = "full", lang = "zh")
asr <- add_labels(asr, label_type = "full", lang = "en")
}
