% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageadjust.R
\name{ageadjust}
\alias{ageadjust}
\title{Calculate the age standardized rate using the direct method.}
\usage{
ageadjust(
  count,
  pop,
  rate = NULL,
  stdpop,
  method = "gamma",
  conf_level = 0.95,
  mp = 1e+05
)
}
\arguments{
\item{count}{The number of cases of a specific disease or condition.}

\item{pop}{The total population of the same group or region
where the disease cases (count) were observed.}

\item{rate}{Disease rate, which is the number of cases (count) per unit
of population (pop).}

\item{stdpop}{Standardized population for age standardization.}

\item{method}{Method used for calculating the age-standardized rate, options
are 'gamma', 'normal', or 'lognormal', default is 'gamma'.}

\item{conf_level}{Confidence level for calculating confidence
intervals, value between 0 and 1, default is 0.95.}

\item{mp}{A multiplier used to scale the calculated rates. Default is 100000.}
}
\value{
Age standardized rate and its confidence interval.
}
\description{
Calculate the age standardized rate using the direct method.
}
\examples{
cases <- c(50, 60, 45, 70)
pop <- c(1000, 1200, 1100, 900)
spop <- c(800, 1000, 1100, 900)
ageadjust(cases, pop, stdpop = spop, mp = 1000)
}
