% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_rate.R
\name{asr_rate}
\alias{asr_rate}
\title{Compute age standardized rate based on provided standard population}
\usage{
asr_rate(
  data,
  std_pop,
  ...,
  event = "fbs",
  pop = "rks",
  agegrp = "agegrp",
  std = c("segi", "china"),
  mp = 1e+05,
  sig = 0.95
)
}
\arguments{
\item{data}{A data frame or tibble that containing case counts, population
in risk, and stratification variables for calculating standardized
rates.}

\item{std_pop}{A data frame or tibble that containing the standard
population, the variable name for age group must be 'agegrp'.}

\item{...}{Variable or variables for stratification when compute
age standard rates.}

\item{event}{A variable within the input data that corresponds to the cases
counts.}

\item{pop}{A variable within the input data that corresponds to the
population in risk in each stratification.}

\item{agegrp}{A variable within the input data that corresponds to the age
categories.}

\item{std}{Specify the standard population structure in the 'std_pop'
data frame used for calculating standardized rates. When calculating
standardized rates for multiple standard populations, specify
std = c("segi", "china").}

\item{mp}{A constant to multiply rates by (e.g. mp=1000 for rates per 1000).}

\item{sig}{The desired level of confidence in computing confidence intervals.
The default is 0.95 for 95 percent CIs.}
}
\value{
A data frame or tibble contains the age standard rates and CIs.
Variables included in the output data frame as follows.
\describe{
\item{...}{Variable or variables for stratification.}
\item{no_cases}{Sum of cases in each stratification.}
\item{cr}{Crude rate in each stratification.}
\item{cr_var}{Variance of the crude rate.}
\item{cr_lower}{Lower limit of the crude rate confidence interval.}
\item{cr_upper}{Upper limit of the crude rate confidence interval.}
\item{asr_china}{Age standardized rate using the china standard pop.
The output variable name would be "asr_" followed by the name of the
standard population, if other standard population was used.}
\item{asr_var_china}{Variance of the age standard rate using china
standard population.}
\item{asr_lower_china}{Lower limit of the age standard rate confidence
interval using china standard population.}
\item{asr_upper_china}{Upper limit of the age standard rate confidence
interval using china standard population.}
}
}
\description{
This function compute the crude rate and age standard rates and their 95\%
confidence intervals(CIs).
}
\references{
\insertRef{fay1997}{canregtools}
}
\author{
\href{https://www.chenq.site}{Qiong Chen},
Email:\href{mailto:chenqiong\@hnccr.org.cn}{chenqiong\@hnccr.org.cn}
}
