% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg.R
\docType{class}
\name{canreg-class}
\alias{canreg-class}
\title{Class of canreg}
\value{
Class of canreg .
}
\description{
Class of canreg
}
\section{Slots}{

\describe{
\item{\code{areacode}}{character.}

\item{\code{name}}{character.}

\item{\code{contains}}{c.}
}}

