% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canregs}
\alias{canregs}
\title{Example Population-Based Cancer Registry Data (\code{canregs})}
\format{
A list of PBCR datasets with class \code{canregs}.
}
\source{
Henan Province Cancer Registry, China.
}
\usage{
canregs
}
\description{
This dataset provides example data from population-based cancer registries (PBCRs),
structured using the \code{canregs} class. It can be used for demonstration, testing,
or development purposes within the \code{canregtools} package.
}
\examples{
data(canregs)
summary(canregs[[1]])
}
\keyword{datasets}
