% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_canreg.R
\name{check_canreg}
\alias{check_canreg}
\alias{check_canreg.canregs}
\alias{check_canreg.canreg}
\title{Check the class of 'canreg'}
\usage{
check_canreg(data, pop_type = "single")

\method{check_canreg}{canregs}(data, pop_type = "single")

\method{check_canreg}{canreg}(data, pop_type = "single")
}
\arguments{
\item{data}{Object of classs 'canreg'}

\item{pop_type}{Population data structure of class 'canreg'}
}
\value{
Logical value
}
\description{
Check the class of 'canreg'
}
