% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_group_vars}
\alias{check_group_vars}
\title{Check the legality of group vars passed by dot dot dot}
\usage{
check_group_vars(x, ..., quiet = TRUE)
}
\arguments{
\item{x}{x}

\item{...}{Dot dot dot parameters}

\item{quiet}{Run the function quietly or not.}
}
\value{
Logical value
}
\description{
Check the legality of group vars passed by dot dot dot
}
