% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_childhood.R
\name{classify_childhood}
\alias{classify_childhood}
\title{Classify the Childhood cancer into ICCC3}
\usage{
classify_childhood(topo, morp, type = "main", lang = "cn")
}
\arguments{
\item{topo}{Topography parts of ICDO3 codes in the format of 'C15.6' or
'C156'.}

\item{morp}{Morphology parts of ICDO3 codes in the format of '8000' or
'M8140'.}

\item{type}{Type of Classification for the output, options are 'main' or
'sub', default is 'main'.}

\item{lang}{Language of the output, options are 'cn' or 'en',
default is 'cn'.}
}
\value{
Factor of ICCC3 classification for childhood cancer.
}
\description{
Classify the Childhood cancer into ICCC3
}
\details{
This function classify the 'topo' and 'morp' parts of ICDO3 codes into
ICCC3 (International Classification of Childhood Cancer, Third edition).
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
fbcases <- data$FBcases
fbcases <- fbcases[fbcases$age <= 14,]
classify_childhood(fbcases$topo, fbcases$morp, lang = "cn")
classify_childhood(fbcases$topo, fbcases$morp, lang = "en")
classify_childhood(fbcases$topo, fbcases$morp, type = "sub", lang = "cn")
classify_childhood(fbcases$topo, fbcases$morp, type = "sub", lang = "en")

}
