% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_icd102.R
\name{classify_icd102}
\alias{classify_icd102}
\title{CConvert ICD10 Codes to Cancer Categories.}
\usage{
classify_icd102(x, cancer_type = "big", as_factor = FALSE)
}
\arguments{
\item{x}{The ICD10 codes of cancer part (C00-C98 and D00-D48) collected
by the Population-Based Cancer Registration (PBCR).}

\item{cancer_type}{Character, type of cancer category ("system","big","small","gco"),
by default, the "big" category was choosed which will categorize the
icd10 code 26 cancer categories.}

\item{as_factor}{logical indicate that the output value as factor.}
}
\value{
Cancer code
}
\description{
Convert ICD10 codes to Cancer Categories according to the specified
category type and language.
}
\examples{
icd10 <- c("C15.2", "C33.4", "C80.9", "C26.2", "C16.3")
classify_icd102(icd10, cancer_type = "big")
}
