% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_canreg.R
\name{count_canreg.canregs}
\alias{count_canreg.canregs}
\title{count the canreg data for class canregs.}
\usage{
\method{count_canreg}{canregs}(x, ...)
}
\arguments{
\item{x}{data of class canregs.}

\item{...}{Others options.}
}
\value{
An object of class fbswicds.
}
\description{
count the canreg data for class canregs.
}
