% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_clean.R
\name{cr_clean}
\alias{cr_clean}
\alias{cr_clean.canregs}
\alias{cr_clean.canreg}
\alias{cr_clean.FBcases}
\alias{cr_clean.SWcases}
\alias{cr_clean.POP}
\title{Clean canreg data.}
\usage{
cr_clean(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{cr_clean}{canregs}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{cr_clean}{canreg}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{cr_clean}{FBcases}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{cr_clean}{SWcases}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{cr_clean}{POP}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)
}
\arguments{
\item{x}{Data of class 'FBcases', 'SWcases' or 'population'.}

\item{age_breaks}{Cut points for age groups. Default is
\code{c(0, 1, seq(5, 85, 5))}.}

\item{label_tail}{Tail of the labels.}

\item{cancer_type}{A character string specifying the classification method
used to categorize ICD-10 codes. This determines how ICD-10 codes are
classified. Options include \code{"big"} (classify ICD-10 codes into 26
cancer categories), \code{"small"} (classify ICD-10 codes into 59 cancer
categories, more specific categories), \code{"system"} (classify ICD-10
codes into organ system), and \code{"gco"} (classify ICD-10 code into
cancer categories same as classification published by the Global
Cancer Observatory). This parameter is only available when the
input data is a vector of ICD-10 codes, or object with class of
\code{'canreg'} or \code{'canregs'}.}
}
\value{
Class 'canreg'.
}
\description{
Clean canreg data.
}
