% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_select.R
\name{cr_select}
\alias{cr_select}
\alias{cr_select.canregs}
\alias{cr_select.asrs}
\alias{cr_select.fbswicds}
\alias{cr_select.summaries}
\title{Select elements from objects of class canregs, fbswicds, or asrs}
\usage{
cr_select(data, ..., index = names(data))

\method{cr_select}{canregs}(data, ..., index = names(data))

\method{cr_select}{asrs}(data, ..., index = names(data))

\method{cr_select}{fbswicds}(data, ..., index = names(data))

\method{cr_select}{summaries}(data, ..., index = names(data))
}
\arguments{
\item{data}{An object of class 'canregs', 'asrs', or 'fbswicds'.}

\item{...}{Conditions or expressions to subset the data within list elements.}

\item{index}{A vector of character strings, logical values, or numeric
indices to select object elements.}
}
\value{
An object with selected elements from the input object that meet the
specified conditions or index. The returned object will have the
same class as the input object.
}
\description{
Select elements from object of class 'canregs', 'fbswicds', or 'asrs' based
on provided index or conditions. The class of the returned object will be the
same with the input object.
}
