% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_rate.R
\name{create_age_rate}
\alias{create_age_rate}
\alias{create_age_rate.canreg}
\alias{create_age_rate.fbswicds}
\alias{create_age_rate.fbswicd}
\title{Create age specific rate}
\usage{
create_age_rate(
  x,
  ...,
  event = fbs,
  cancer_type = "big",
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)

\method{create_age_rate}{canreg}(x, ..., cancer_type = "big")

\method{create_age_rate}{fbswicds}(
  x,
  ...,
  event = fbs,
  collapse = TRUE,
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)

\method{create_age_rate}{fbswicd}(
  x,
  ...,
  event = fbs,
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)
}
\arguments{
\item{x}{The input data. For \code{create_age_rate}, this could be an object of
class \code{canreg} or \code{fbswicd}.}

\item{...}{Variable name(s) used for stratification.}

\item{event}{Event used for calculating age-specific rates.}

\item{cancer_type}{Method used to classify ICD10.}

\item{format}{Format of the output data frame, either "long" or "wide".}

\item{mp}{Correction factor, default is 100000.}

\item{decimal}{Number of decimal places for rounding, default is 6.}

\item{show_pop}{Logical value whether output population or not.}

\item{collapse}{Whether the output collapse to tibble.}
}
\value{
A data frame of age-specific rates.
}
\description{
Creates a data frame of age-specific rates based on the class of data
(\code{fbswicd} or \code{canreg}).
}
