% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_asr.R
\name{create_asr.canreg}
\alias{create_asr.canreg}
\title{Calculate age-standardized rate.}
\usage{
\method{create_asr}{canreg}(x, ..., type = "big", lang = "cn")
}
\arguments{
\item{x}{Object with class of canreg.}

\item{...}{Parameters.}

\item{type}{Classification type of cancer site, options are "big", "small"
or "system", default is "big".}

\item{lang}{Language of the output result, options are "cn" for chinese, and
"en" for english, default is "cn".}
}
\value{
class of asr.
}
\description{
Calculate age-standardized rate.
}
