% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quality.R
\name{create_quality}
\alias{create_quality}
\alias{create_quality.canreg}
\alias{create_quality.canregs}
\alias{create_quality.fbswicds}
\alias{create_quality.fbswicd}
\title{Create quality index from object of class 'fbswicd'.}
\usage{
create_quality(x, ..., decimal = 2)

\method{create_quality}{canreg}(x, ..., cancer_type = "big")

\method{create_quality}{canregs}(x, ..., cancer_type = "big", collapse = TRUE)

\method{create_quality}{fbswicds}(x, ..., decimal = 2, collapse = TRUE)

\method{create_quality}{fbswicd}(x, ..., decimal = 2)
}
\arguments{
\item{x}{An object of class 'fbswicd'.}

\item{...}{Variables used for stratification.}

\item{decimal}{Decimal digits.}

\item{cancer_type}{Method used to classify ICD10.}

\item{collapse}{Whether the output collapse to tibble.}
}
\value{
A data frame of cancer registration quality data.
}
\description{
Create quality index from object of class 'fbswicd'.
}
