% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ratio.R
\name{create_ratio}
\alias{create_ratio}
\alias{create_ratio.fbswicd}
\title{Title}
\usage{
create_ratio(x, ...)

\method{create_ratio}{fbswicd}(x, ...)
}
\arguments{
\item{x}{Data of class 'fbswicd'.}

\item{...}{Stratification variables.}
}
\value{
A data frame.
}
\description{
Title
}
