% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\alias{create_report.canregs}
\alias{create_report.canreg}
\title{Create report from report templates}
\usage{
create_report(
  data,
  template = "annual",
  title = "Cancer Registry Report",
  output_format = "html_document",
  output_dir = NULL,
  ...
)

\method{create_report}{canregs}(
  data,
  template = "annual",
  title = "Cancer Registry Report",
  output_format = "html_document",
  output_dir = NULL,
  ...
)

\method{create_report}{canreg}(
  data,
  template = "annual",
  title = "Cancer Registry Report",
  output_format = "html_document",
  output_dir = NULL,
  ...
)
}
\arguments{
\item{data}{Object with class of 'canreg' or 'fbswicd'.}

\item{template}{Name of the template used to render report, options are
"annual", "quality", or "CI5", default is "annual".}

\item{title}{Title of the report to be rendered, default is "Cancer Registry
Report".}

\item{output_format}{File type of the render result, options are
"html_document", "word_document" or "pdf_document", default is
"html_document".}

\item{output_dir}{Directory of the rendered report.}

\item{...}{Parameters in \code{rmarkdown::render}.}
}
\description{
Create report from report templates
}
