% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheet.R
\name{create_sheet}
\alias{create_sheet}
\alias{create_sheet.canreg}
\alias{create_sheet.fbswicds}
\alias{create_sheet.fbswicd}
\title{Create sheet of cancer statistics}
\usage{
create_sheet(x, sheet_type = "incidence", cancer_type = "big")

\method{create_sheet}{canreg}(x, sheet_type = "incidence", cancer_type = "big")

\method{create_sheet}{fbswicds}(x, sheet_type = "incidence", cancer_type = "big")

\method{create_sheet}{fbswicd}(x, sheet_type = "incidence", cancer_type = "big")
}
\arguments{
\item{x}{data with class of 'fbswicd' or 'canreg'.}

\item{sheet_type}{Character value indicate type of sheet, options are
'incidence', 'mortality', 'quality', 'population'.}

\item{cancer_type}{Type of classification method of ICD10, options are 'big',
'small', 'system', or 'gco', default is "big".}
}
\value{
A data frame or tibble contains the age standard rates and CIs.
}
\description{
Create sheet of cancer statistics
}
