% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{data}
\alias{data}
\title{data}
\arguments{
\item{x}{The input data, object with class of \code{'fbswicd'}, \code{'fbswicds'},
\code{'canreg'}, or \code{'canregs'}.}
}
\description{
data
}
\keyword{internal}
