% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_barchart.R
\name{draw_bar}
\alias{draw_bar}
\title{Draw bar chart.}
\usage{
draw_bar(
  values,
  cates = NULL,
  axis = NULL,
  groups = length(values),
  bar_side = 1,
  bar_way = 1,
  gap = ifelse(side_label == 1, 0.6, 0.1),
  csize = 0.8,
  space = 0.9,
  adj = -0.01,
  gl = TRUE,
  label = c("Male", "Female"),
  cols = c("#006400", "#b32134"),
  side_label = 1,
  legend = FALSE,
  legend_label = NULL,
  dens = c(-1, -1),
  overlay = FALSE,
  ...
)
}
\arguments{
\item{values}{List of values to be draw.}

\item{cates}{Categories for the groups.}

\item{axis}{Axis values.}

\item{groups}{Groups of the bar chart, default is length of the values list.}

\item{bar_side}{Orientation of the bar chart: 1 for left and 2 for right.}

\item{bar_way}{Draw the bar chart unidirectionally or bidirectionally, 1 for
unidirectionally and 2 for bidirectionally.}

\item{gap}{The width of center gap. Default is 0.1.}

\item{csize}{The font size to write the labels on sides. Default is 1.}

\item{space}{Space between bars.}

\item{adj}{The vertical adjustment factor for the labels of age classes.
Default is -0.01.}

\item{gl}{Logical value to draw the vertical dotted lines. Default is TRUE.}

\item{label}{Labels for the two sides.}

\item{cols}{The colors of the bars.}

\item{side_label}{Draw the side label on the left or right. 1 for left and
2 for right.}

\item{legend}{Logical value indicates plot legend or not.}

\item{legend_label}{Character for the legend label.}

\item{dens}{The density of hatching lines (/inch).}

\item{overlay}{Logical value, TRUE for overlay plot, FALSE for not.}

\item{...}{Other options.}
}
\value{
A bar chart.
}
\description{
Draw bar chart.
}
