% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_barchart.R
\name{draw_dumbbell}
\alias{draw_dumbbell}
\title{Plot dumbbell chart.}
\usage{
draw_dumbbell(
  data,
  x = NULL,
  y1 = NULL,
  y2 = NULL,
  topn = 20,
  sort = "insc",
  legend = NULL,
  cols = c("#006400", "#b32134"),
  main = ""
)
}
\arguments{
\item{data}{A data frame contains data to be plotted.}

\item{x}{A category variable in data.}

\item{y1}{Variable indicate start point.}

\item{y2}{Variable indicate end point.}

\item{topn}{Top n values to be plotted.}

\item{sort}{Sort options.}

\item{legend}{Legends.}

\item{cols}{Colors of the start and end points.}

\item{main}{Main title of the plot.}
}
\value{
A dumbbell plot.
}
\description{
Plot dumbbell chart.
}
