% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_linechart.R
\name{draw_line}
\alias{draw_line}
\title{Plot line chart}
\usage{
draw_line(
  data,
  x_var,
  y_var,
  group_var,
  y_axis = NULL,
  x_axis = NULL,
  col = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{data}{The input data frame.}

\item{x_var}{The variable for the x-axis.}

\item{y_var}{The variable for the y-axis.}

\item{group_var}{The variable used to group the data for multiple lines.}

\item{y_axis}{Optional. The custom Y-axis tick values. If not provided, the
function generates them.}

\item{x_axis}{Optional. The custom X-axis tick values. If not provided, the
function generates them.}

\item{col}{Optional. The colors for the lines. If not provided, the function
uses default colors.}

\item{main}{Optional. The title of the plot.}

\item{...}{Other parameters.}
}
\value{
Line plot.
}
\description{
Plot line chart
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
fbsw <- count_canreg(data, lang = "en")
age <- create_age_rate(fbsw, sex)
draw_line(age, agegrp, rate, sex)
}
