% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_linechart.R
\name{draw_linechart}
\alias{draw_linechart}
\title{Draw a Custom Line Chart Using Base R}
\usage{
draw_linechart(
  data,
  x,
  y,
  group = NULL,
  facet = NULL,
  grid = c(1, 1),
  x_axis = NULL,
  y_axis = NULL,
  x_label = NULL,
  y_label = NULL,
  axis_title = c("Age (years)", "Age specific rate"),
  cols = NULL,
  palette = "Peach",
  line_type = "l",
  lwd = 2,
  adj = 0.02,
  srt = 60,
  main = NULL,
  sub = NULL,
  mar = c(1, 0, 1, 0),
  add = TRUE,
  offset = 0.01,
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables to plot.}

\item{x, y}{Bare column names for the x and y axis variables.}

\item{group}{Optional bare column name used to group and color lines.}

\item{facet}{Optional bare column name used for faceting. If provided,
the data will be split by this variable and plotted in a multi-panel
layout.}

\item{grid}{A vector of length 2 specifying number of rows and columns for
facets. Default is c(1, 1).}

\item{x_axis}{Optional numeric vector specifying x-axis tick locations.}

\item{y_axis}{Optional numeric vector specifying y-axis tick locations.}

\item{x_label, y_label}{Optional labels for x and y axis ticks. If \code{NULL},
defaults are used.}

\item{axis_title}{Character vector of length 2 giving the axis titles:
c("x axis label", "y axis label").}

\item{cols}{Character vector of line colors. Defaults to c("darkgreen",
"darkred", "gray").}

\item{palette}{Character, palette name indicate group of colors.}

\item{line_type}{1-character string giving the type of plot desired. The
following values are possible, for details, see plot: "p" for points,
"l" for lines, "b" for both points and lines, "c" for empty points
joined by lines, "o" for overplotted points and lines, "s" and "S"
for stair steps and "h" for histogram-like vertical lines. Finally,
"n" does not produce any points or lines.}

\item{lwd}{Line width. Default is 2.}

\item{adj}{Adjustment for axis text placement. Default is 0.02.}

\item{srt}{String rotation angle for x-axis labels. Default is 90 degrees.}

\item{main, sub}{Main title and subtitle of the plot.}

\item{mar}{Margin of the sub plot.}

\item{add}{Logical. If \code{TRUE}, restores original graphics parameters after
plotting.}

\item{offset}{Axis offset used for spacing ticks. Default is 0.01.}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function draws a line chart from a data frame, optionally grouped by a
categorical variable. It uses base R graphics and supports custom axis
ticks, labels, and styles.
}
\examples{
data("canregs")
fbsw <- count_canreg(canregs[[1]], label_tail="yrs")
agerate <- create_age_rate(fbsw, year, sex)
agerate <- add_labels(agerate, lang = "en")
draw_linechart(agerate, agegrp, rate, sex)
agerate <- create_age_rate(fbsw, year, sex, cancer)
agerate <- add_labels(agerate, lang = "en")
agerate <- dplyr::filter(agerate, cancer \%in\% as.character(c(103:106)))
draw_linechart(agerate, agegrp, rate, sex, cancer, grid = c(2, 2))
}
