% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pyramid.R
\name{draw_pyramid.fbswicd}
\alias{draw_pyramid.fbswicd}
\title{Draw pyramids from object of fbswicd.}
\usage{
\method{draw_pyramid}{fbswicd}(x, ..., grid = NULL)
}
\arguments{
\item{x}{Object of class 'fbswicd'.}

\item{...}{Parameters.}

\item{grid}{grids.}
}
\value{
Pyramid plots.
}
\description{
Draw pyramids from object of fbswicd.
}
