% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pyramid.R
\name{draw_pyramid}
\alias{draw_pyramid}
\title{Draw population pyramids.}
\usage{
draw_pyramid(
  x,
  show_value = TRUE,
  show_prop = TRUE,
  left_axis = NULL,
  right_axis = NULL,
  axis_fm = "g",
  axis_bm = "",
  axis_bi = 3,
  cgap = 0.3,
  cstep = 1,
  csize = 1,
  labs = c("Males", "Ages", "Females"),
  gl = TRUE,
  cadj = 0,
  cols = c("#006400", "#b32134"),
  dens = c(-1, -1),
  main = "",
  grid = c(2, 3)
)
}
\arguments{
\item{x}{Class of 'canreg' or 'fbswicd'.}

\item{show_value}{Logical value to draw the population numbers. Default is
TRUE.}

\item{show_prop}{Logical value to draw the pyramid using proportion.
Default is TRUE.}

\item{left_axis}{A vector of axis for left pyramid. If missing, automatically
given using pretty().}

\item{right_axis}{A vector of axis for right pyramid.
If missing, left_axis is used.}

\item{axis_fm}{A format code of formatC for plotting axis.
If missing, "g" is used.}

\item{axis_bm}{A big.mark of formatC for plotting axis.
If missing, none.}

\item{axis_bi}{A big.interval number of formatC for plotting axis.
Default is 3.}

\item{cgap}{The width of center gap (as ratio to each panel) to draw
age-class. Default is 0.3.}

\item{cstep}{The interval to write the labels of age classes.
Default is 1.}

\item{csize}{The font size factor to write the labels of age classes.
Default is 1.}

\item{labs}{The label of the left, center, and right pyramid.
Default is c("Males", "Ages", "Females").}

\item{gl}{Logical value to draw the vertical dotted lines. Default is TRUE.}

\item{cadj}{The vertical adjustment factor for the labels of age classes.
Default is -0.03.}

\item{cols}{The color of the left and right pyramid.
Default is c("lightblue", "pink").}

\item{dens}{The density of hatching lines (/inch) for left and right pyramid.
Default is c(-1, -1), when the pyramid will be filled.}

\item{main}{The main title of the pyramid.}

\item{grid}{Grids.}
}
\value{
Pyramids plot.
}
\description{
Draw population pyramids.
}
