% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_deal.R
\name{drop_others}
\alias{drop_others}
\title{Drop other cancer in cancer categories.}
\usage{
drop_others(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame.
}
\description{
Drop other cancer in cancer categories.
}
