% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{dulbar_chart}
\alias{dulbar_chart}
\title{Plot dual bar chart.}
\usage{
dulbar_chart(
  data,
  topn = 10,
  gap = 0.1,
  space = 0.9,
  adj = -0.01,
  csize = 1,
  gl = TRUE,
  label = c("Male", "Female"),
  legend = c("Crude rate", "ASR rate"),
  cols = c("#006400", "#b32134"),
  dens = c(-1, -1),
  main = "",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{topn}{Top n values to show.}

\item{gap}{The width of center gap. Default is 0.1.}

\item{space}{Space between bars.}

\item{adj}{The vertical adjustment factor for the labels of age classes.
Default is -0.01.}

\item{csize}{The font size to write the labels on sides. Default is 1.}

\item{gl}{Logical value to draw the vertical dotted lines. Default is TRUE.}

\item{label}{Labels for the two sides.}

\item{legend}{Legends for the bars.}

\item{cols}{The colors of the bars.}

\item{dens}{The density of hatching lines (/inch).}

\item{main}{The main title for the dulbar.}

\item{...}{Other options.}
}
\value{
A dulbar chart.
}
\description{
Plot dual bar chart.
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
rate <- asr_rate(data, icd_cat, event = sws)
dulbar_chart(rate[,c(1, 3, 6, 9, 10)])
}
