% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esti_pop.R
\name{esti_pop}
\alias{esti_pop}
\title{Estimating population structure using interpolation method.}
\usage{
esti_pop(pop1, pop2, period)
}
\arguments{
\item{pop1}{Population or population proportion in each age group for the
start year.}

\item{pop2}{Population or population proportion in each age group for the
end year.}

\item{period}{Vector contain the start year and end year value.}
}
\value{
A data frame contain the estimated population proportion in each
year during the period with each year in one column and each age
group in one row.
}
\description{
Estimating population structure using interpolation method.
}
\examples{

pop1 <- c(
  59546, 294129, 472511, 552549, 821119, 996436, 805635, 1004506,
  989357, 1056612, 986559, 792270, 544544, 452297, 473579, 350802,
  212614, 109598, 61990
)
pop2 <- c(
  75641, 377276, 327116, 380338, 539034, 1158852, 1152329, 881443,
  903484, 1011164, 1238871, 1137832, 1022787, 645441, 464777,
  482941, 406144, 227977, 144526
)
esti_pop(pop1, pop2, c(2000, 2010))

}
