% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{event}
\alias{event}
\title{event}
\arguments{
\item{event}{A variable used to specify the type of calculation, options are
fbs or sws, fbs for cancer incidence, and sws for cancer mortality.}
}
\description{
event
}
\keyword{internal}
