% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{expand_age_pop}
\alias{expand_age_pop}
\title{Expand age groups population.}
\usage{
expand_age_pop(x, method = "linear")
}
\arguments{
\item{x}{Vector, population for each age group.}

\item{method}{Method for expanding, options are 'linear', 'constant',
'periodic', or 'natural'. Default is 'linear'.}
}
\value{
Vector contain the expanded age population.
}
\description{
Expand age groups population.
}
\examples{
ages <- c(
  5053, 17743, 25541, 32509, 30530, 34806, 36846, 38691, 40056,
  39252, 37349, 30507, 26363, 21684, 15362, 11725, 7461, 3260, 915)
expand_age_pop(ages)
}
