% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{expand_lifetable}
\alias{expand_lifetable}
\title{Expand five-year(abridged) life table to one-year(complete) life table.}
\usage{
expand_lifetable(lx)
}
\arguments{
\item{lx}{A vector contains the total number of survivors lx at age x
which was obtained from the abridged life table.}
}
\value{
List contains two vectors fitlx and fitmx.
}
\description{
Expand five-year(abridged) life table to one-year(complete) life table.
}
\examples{
lx <- c(
  100000, 99498.39, 99294.62, 99173.88, 99047.59, 98840.46,
  98521.16, 98161.25, 97636.99, 96900.13, 95718.96, 93930.91,
  91463.21, 87131.41, 80525.02, 70907.59, 58090.75, 41630.48,
  24019.33
)
lx <- lx / 100000
expand_lifetable(lx)
}
