% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{expand_lx}
\alias{expand_lx}
\title{Expand a five-year abridged life table to a complete life table}
\usage{
expand_lx(lx, sage = c(0, 1, seq(5, 85, 5)), max_age = 100)
}
\arguments{
\item{lx}{A numeric vector representing the number of survivors (\eqn{l_x})
at the beginning of each age interval in the abridged life table. The
vector should correspond to age intervals provided in \code{sage}.}

\item{sage}{Numeric vector of starting ages for the age groups,
default is \code{c(0, 1, seq(5, 85, 5))}.}

\item{max_age}{Numeric value specifying the max age of the output estimated
lx(fitlx) and mx (fitmx).}
}
\value{
A list containing:
\describe{
\item{fitlx}{A numeric vector representing the estimated
number of survivors at each single year of age from 0 to 100.}
\item{fitmx}{A numeric vector representing the estimated
central death rates (\eqn{m_x}) for each single year of
age from 0 to 100.}
}
}
\description{
\code{expand_lx()} transforms a five-year abridged life table into a
one-year complete life table using the Elandt–Johnson method.
}
\examples{
# Example abridged life table data (normalized to a radix of 1)
lx <- c(
100000, 99498.39, 99294.62, 99173.88, 99047.59, 98840.46,
98521.16, 98161.25, 97636.99, 96900.13, 95718.96, 93930.91,
91463.21, 87131.41, 80525.02, 70907.59, 58090.75, 41630.48,
24019.33
)
lx <- lx / 100000
expand_lx(lx)
}
\references{
Baili, P., Micheli, A., Montanari, A., & Capocaccia, R. (2005).
Comparison of Four Methods for Estimating Complete Life Tables
from Abridged Life Tables Using Mortality Data Supplied to EUROCARE-3.
\emph{Mathematical Population Studies}, 12(4), 183–198.
https://doi.org/10.1080/08898480500301751
}
