% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg.R
\docType{class}
\name{FBcases-class}
\alias{FBcases-class}
\title{Class of FBcases}
\value{
class of FBcases.
}
\description{
Class of FBcases
}
\section{Slots}{

\describe{
\item{\code{registr}}{character.}

\item{\code{id}}{character.}

\item{\code{name}}{character.}

\item{\code{sex}}{factor.}

\item{\code{age}}{numeric.}

\item{\code{birthda}}{Date.}

\item{\code{topo}}{character.}

\item{\code{morp}}{character.}

\item{\code{beha}}{character.}

\item{\code{grad}}{character.}

\item{\code{basi}}{character.}

\item{\code{icd10}}{character.}

\item{\code{inciden}}{Date.}

\item{\code{dlc}}{Date.}

\item{\code{status}}{integer.}

\item{\code{deathda}}{Date.}

\item{\code{caus}}{integer.}

\item{\code{areacode}}{character.}
}}

