% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbswicd.R
\name{fbswicd}
\alias{fbswicd}
\title{Count cancer incident cases and deaths in each year, sex, cancer sites
and generate a data frame.}
\usage{
fbswicd(
  fb,
  sw,
  pop,
  year = year,
  sex = sex,
  agegrp = agegrp,
  icd_cat = icd_cat,
  basi = basi
)
}
\arguments{
\item{fb}{A data frame that generated by prep_regdata function.}

\item{sw}{A data frame that generated by prep_regdata function.}

\item{pop}{A data frame that generated by prep_regdata function.}

\item{year}{Variable name for year.}

\item{sex}{Variable name for gender.}

\item{agegrp}{Variable name for age categories.}

\item{icd_cat}{Variable name for cancer categories.}

\item{basi}{Variable name for diagnostic basi for cancer cases.}
}
\value{
A data frame that contain the counts of cancer cases and deaths.
}
\description{
Count cancer incident cases and deaths in each year, sex, cancer sites
and generate a data frame.
}
\examples{
library(canregtools)
library(readxl)
file <- system.file("extdata", "411721.xls", package = "canregtools")
FB <- read_excel(file, sheet="FB")
SW <- read_excel(file, sheet="SW")
POP <- read_excel(file, sheet="POP")
fb <- prep_regdata(FB, icd_type = "big")
sw <- prep_regdata(SW, icd_type = "big")
pop <- prep_regdata(POP)
fbsw <- fbswicd(fb, sw, pop)
}
