% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop.R
\name{get_pop}
\alias{get_pop}
\alias{get_pop.canreg}
\alias{get_pop.canregs}
\alias{get_pop.fbswicd}
\alias{get_pop.fbswicds}
\title{Subset or summarize population data}
\usage{
get_pop(data, sum_by = NULL, collapse = FALSE)

\method{get_pop}{canreg}(data, sum_by = NULL, collapse = FALSE)

\method{get_pop}{canregs}(data, sum_by = NULL, collapse = FALSE)

\method{get_pop}{fbswicd}(data, sum_by = NULL, collapse = FALSE)

\method{get_pop}{fbswicds}(data, sum_by = NULL, collapse = FALSE)
}
\arguments{
\item{data}{An object of class \code{canreg}, \code{canregs}, \code{fbswicd}, or \code{fbswicds}.}

\item{sum_by}{Character vector of grouping variables to summarize population.}

\item{collapse}{Logical, if \code{TRUE}, collapses the list of results into a single data frame (only for list-type objects).}
}
\value{
A data frame or a list of data frames depending on \code{collapse}.
}
\description{
Extracts population data from canreg-style objects and optionally summarizes it
by specified grouping variables.
}
