% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_select.R
\name{has_rows}
\alias{has_rows}
\title{Check if the data frame has rows after filtering}
\usage{
has_rows(df, ...)
}
\arguments{
\item{df}{A data frame to check for row count after applying conditions.}

\item{...}{Conditions for filtering the data frame.}
}
\value{
Logical: TRUE if the data frame has rows, FALSE otherwise.
}
\description{
Check if the data frame has rows after filtering
}
