% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdo3to10.R
\name{icdo3to10}
\alias{icdo3to10}
\title{Convert ICDO3 codes to ICD10 code.}
\usage{
icdo3to10(topo, morp, beha)
}
\arguments{
\item{topo}{A vector containing ICDO3 topography codes, it can be
in the format of either "C15.1" or "C151".}

\item{morp}{A vector containing ICDO3 morphology codes, it could be in the
format of "M8170" or "8170".}

\item{beha}{A vector containing ICDO3 behavior codes, it could be one of the
values c("0", "1", "2", "3", "6").}
}
\value{
A vector containing ICD10 code.
}
\description{
Convert the ICDO3 codes to ICD10 code based on Topography, Morphology, and
Behavior value.
}
\details{
ICD-O-3 (International Classification of Diseases for Oncology, 3rd Edition)
and ICD-10 (International Classification of Diseases, 10th Edition) are two
different coding systems used for disease classification. ICD-10 is a widely
used coding system that can describe various diseases, including tumor
diseases. On the other hand, ICD-O-3 is specifically designed for the coding
of tumor diseases. Promoting and using ICD-O-3 and ICD-10 helps
in standardizing and unifying the classification and coding of tumor,
facilitating global tumor epidemiology research, statistical analysis,
and clinical management. ICD-O-3 consists of three main components:
Topography, Morphology, and Behavior. The Topography code uses a four-digit
number to describe the specific location where the tumor occurs. The
Morphology code uses a three-digit number to represent the morphology or
histology type of the tumor. The Behavior code describes the behavior
characteristics of the tumor. With this program, you can convert ICD-O-3
codes to ICD-10 codes, making it convenient for research and
clinical practice.
}
\examples{

topo <- c("C001","C15.1","C349","509","C421","C774")
morp <- c("")
}
