% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{load_canreg}
\alias{load_canreg}
\title{Load Example Data Set}
\usage{
load_canreg()
}
\value{
A data frame or tibble containing the example data.
}
\description{
This function loads an example data set from the 'canregtools' package.
The data set is an Excel file included with the package.
}
\examples{
data <- load_canreg()

}
