% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_vars.R
\name{ls_attrs}
\alias{ls_attrs}
\title{list built-in attributes name used in \code{cr_reframe()}}
\usage{
ls_attrs()
}
\value{
A character vector.
}
\description{
list built-in attributes name used in \code{cr_reframe()}
}
\examples{
ls_attrs()

}
