% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_reframe_vars}
\alias{ls_reframe_vars}
\title{List names of variables could be used cr_reframe function.}
\usage{
ls_reframe_vars()
}
\value{
A \code{tibble} with two columns:
\itemize{
\item \code{Vars}: The name of the variable (e.g., "fbs").
\item \code{Description}: A detailed description of the variable.
}
}
\description{
List names of variables could be used cr_reframe function.
}
\examples{
ls_reframe_vars()
}
