% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ls_std_vars}
\alias{ls_std_vars}
\title{list all standard population names available}
\usage{
ls_std_vars()
}
\value{
Character vector contain names of standard population.
}
\description{
list all standard population names available
}
