% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mp_decimal}
\alias{mp_decimal}
\title{mp decimal}
\arguments{
\item{mp}{A constant to multiply rates by (e.g. mp=1000 for rates per 1000).}

\item{decimal}{This parameter specifies the number of decimal places to
round the results. The default is 2, which means rates will
be rounded to two decimal places.}
}
\description{
mp decimal
}
\keyword{internal}
