% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg_report.R
\name{new_canreg_report}
\alias{new_canreg_report}
\title{Construct a \code{canreg_report} object}
\usage{
new_canreg_report(meta, source = NULL, pop, quality, inci, mort)
}
\arguments{
\item{meta}{Metadata list.}

\item{source}{Optional intermediate source objects.}

\item{pop, quality, inci, mort}{Report component lists.}
}
\value{
An object of class \code{canreg_report}.
}
\description{
Construct a \code{canreg_report} object
}
