% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg.R
\docType{class}
\name{population-class}
\alias{population-class}
\title{Class of population}
\value{
class of population.
}
\description{
Class of population
}
\section{Slots}{

\describe{
\item{\code{year}}{integer.}

\item{\code{sex}}{factor.}

\item{\code{agegrp}}{factor.}

\item{\code{rks}}{numeric.}
}}

