% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pre_deal_fbswicd}
\alias{pre_deal_fbswicd}
\title{Prepare the fbswicd data before other calculation}
\usage{
pre_deal_fbswicd(x, ...)
}
\arguments{
\item{x}{Object with class of fbswicd}

\item{...}{Dot dot dot}
}
\value{
fbswicd data removed combined stats.
}
\description{
Prepare the fbswicd data before other calculation
}
