% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pre_deal_pop}
\alias{pre_deal_pop}
\title{Reframe the population data according to vars passed by dot dot dot.}
\usage{
pre_deal_pop(x, ...)
}
\arguments{
\item{x}{Population data.}

\item{...}{Dot dot dot.}
}
\value{
Reframed population data.
}
\description{
Reframe the population data according to vars passed by dot dot dot.
}
