% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_regdata.R
\name{prep_regdata}
\alias{prep_regdata}
\title{Prepare the population based cancer registration data}
\usage{
prep_regdata(data, icd_type = "big")
}
\arguments{
\item{data}{Data frame.}

\item{icd_type}{Specify the classification method for ICD-10.}
}
\value{
Data frame includes newly generated variables such as year of
incidence or death, age groups, cancer classifications, etc.
}
\description{
Organize data (FB, SW, and POP) for further analysis using fbswicd function.
}
\examples{
library(canregtools)
library(readxl)
file <- system.file("extdata", "411721.xls", package = "canregtools")
FB <- read_excel(file, sheet="FB")
fb <- prep_regdata(FB, icd_type = "big")

}
\author{
\href{https://www.chenq.site}{Qiong Chen}, Email: chenqiong@hnccr.org.cn
}
