% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{Plot population pyramid}
\usage{
pyramid(
  data,
  Show.Values = TRUE,
  Show.Proportion = TRUE,
  Laxis = NULL,
  Raxis = NULL,
  AxisFM = "g",
  AxisBM = "",
  AxisBI = 3,
  Cgap = 0.3,
  Cstep = 1,
  Csize = 1,
  Llab = "Males",
  Rlab = "Females",
  Clab = "Ages",
  GL = TRUE,
  Cadj = 0,
  Lcol = "lightblue",
  Rcol = "pink",
  Ldens = -1,
  Rdens = -1,
  main = "",
  ...
)
}
\arguments{
\item{data}{A data.frame including left pyramid numbers in the 1st column
and and right pyramid numbers in the 2nd column, where the numbers
of males in each age-class are usually given to left numbers and
those of females are to right numbers. If the data.frame includes 3rd column,
it is used as age-class labels, otherwise the row.names(data) is used as
age-class labels.}

\item{Show.Values}{Logical value to draw the population numbers. Default is
TRUE.}

\item{Show.Proportion}{Logical value to draw the pyramid using proportion.
Default is TRUE.}

\item{Laxis}{A vector of axis for left pyramid. If missing, automatically
given using pretty().}

\item{Raxis}{A vector of axis for right pyramid. If missing, Laxis is used.}

\item{AxisFM}{A format code of formatC for plotting axis.
If missing, "g" is used.}

\item{AxisBM}{A big.mark of formatC for plotting axis.
If missing, none.}

\item{AxisBI}{A big.interval number of formatC for plotting axis.
Default is 3.}

\item{Cgap}{The width of center gap (as ratio to each panel) to draw
age-class. Default is 0.3.}

\item{Cstep}{The interval to write the labels of age classes.
Default is 1.}

\item{Csize}{The font size factor to write the labels of age classes.
Default is 1.}

\item{Llab}{The label of the left pyramid. Default is "Males".}

\item{Rlab}{The label of the right pyramid. Default is "Females".}

\item{Clab}{The label of the center age-class. Default is "Ages".}

\item{GL}{Logical value to draw the vertical dotted lines. Default is TRUE.}

\item{Cadj}{The vertical adjustment factor for the labels of age classes.
Default is -0.03.}

\item{Lcol}{The color of the left pyramid. Default is "Cyan".}

\item{Rcol}{The color of the right pyramid. Default is "Pink".}

\item{Ldens}{The density of hatching lines (/inch) for left pyramid.
Default is -1, when the pyramid will be filled.}

\item{Rdens}{The density of hatching lines (/inch) for right pyramid.
Default is -1, when the pyramid will be filled.}

\item{main}{The main title of the pyramid.}

\item{...}{Other options.}
}
\value{
A population pyramid plot.
}
\description{
Plot population pyramid
}
\examples{
library(canregtools)
library(readxl)
}
