% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_areacode.R
\name{read_areacode}
\alias{read_areacode}
\title{Get area codes from National Bureau of Statistics in China.}
\usage{
read_areacode(
  prov = "41",
  year = "2023",
  level = "county",
  cache_dir = "~/.cache_areacodes",
  cache_refresh = F
)
}
\arguments{
\item{prov}{Province code.}

\item{year}{Year}

\item{level}{Level of returned data, options are 'city', 'county', 'xiang',
'cun', default is 'county'.}

\item{cache_dir}{the directory of cache areacodes.}

\item{cache_refresh}{logical value that refresh cache or not.}
}
\value{
A data frame contain area codes and their labels.
}
\description{
Get area codes from National Bureau of Statistics in China.
}
