% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_canreg.R
\name{read_canreg}
\alias{read_canreg}
\title{Read cancer registration data}
\usage{
read_canreg(
  x,
  pop_type = "long",
  age_var = "agegroup",
  pop_var = "popu",
  death_var = "death"
)
}
\arguments{
\item{x}{A path to an Excel file, a character vector of file paths, or
a directory containing Excel files.}

\item{pop_type}{A character string specifying the format of the population
sheet. Must be either \code{"long"} (default) or \code{"wide"}.}

\item{age_var}{Name of the age group column (used only when
\code{pop_type = "long"}). Default is \code{"agegroup"}.}

\item{pop_var}{Name of the population count column (used only when
\code{pop_type = "long"}). Default is \code{"popu"}.}

\item{death_var}{Name of the death count column (used only when
\code{pop_type = "long"}). If not present in the sheet, a column with 0s will
be added. Default is \code{"death"}.}
}
\value{
A \code{canreg} object (a list with components: \code{areacode}, \code{FBcases},
\code{SWcases}, \code{POP}) if a single file is read. A named list of such objects
with class \code{"canregs"} if multiple files or a directory is provided.
}
\description{
Reads cancer registry data from one or more Excel files or from
a directory containing Excel files. It extracts case and population data and
returns objects of class \code{"canreg"} or a list of such objects (\code{"canregs"}).
}
\examples{
\dontrun{
file_address <- "410302.xlsx"
canreg <- read_canreg(file_address, pop_type = "long")
}

}
