% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{read_map}
\alias{read_map}
\title{Get map of china from dataV.}
\usage{
read_map(
  x,
  level = "province",
  bound = "county",
  cache_dir = "~/.cache_map",
  cache_refresh = FALSE
)
}
\arguments{
\item{x}{Area codes in China, such as "410102".}

\item{level}{Top level of the returned map.}

\item{bound}{Boundary of the returned map.}

\item{cache_dir}{Cache directory of the returned map.}

\item{cache_refresh}{Logical value, TRUE or FALSE rerfresh the stored cached
map or not.}
}
\value{
Map data of sf format.
}
\description{
Get map of china from dataV.
}
\examples{
henan <- read_map("410102", level = "province", bound = "county")
class(henan)
}
