% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe_fbswicd.R
\name{reframe_fbswicd.canregs}
\alias{reframe_fbswicd.canregs}
\title{Reframe the object of classs canregs.}
\usage{
\method{reframe_fbswicd}{canregs}(x, ..., type = "big", lang = "cn")
}
\arguments{
\item{x}{Object of class canregs.}

\item{...}{Stratification variables used to reframe.}

\item{type}{Method for classify cancer sites.}

\item{lang}{Character value for specify the language used. options are 'cn'
for Chinese, or 'en' for English.}
}
\value{
Object of class fbswicd.
}
\description{
Reframe the object of classs canregs.
}
