% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe_fbswicd.R
\name{reframe_fbswicd}
\alias{reframe_fbswicd}
\alias{reframe_fbswicd.canregs}
\alias{reframe_fbswicd.fbswicds}
\title{Reframe data of class fbswicds}
\usage{
reframe_fbswicd(x, ...)

\method{reframe_fbswicd}{canregs}(x, ..., cancer_type = "big")

\method{reframe_fbswicd}{fbswicds}(x, ...)
}
\arguments{
\item{x}{Object of class fbswicds.}

\item{...}{Stratification variables used to reframe.}

\item{cancer_type}{Method for classify cancer sites.}
}
\value{
Merged fbswicd.
}
\description{
Reframe data of class fbswicds
}
