% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_registry.R
\name{show_registry}
\alias{show_registry}
\title{Get the areacodes of the registry}
\usage{
show_registry(regi_type = c(1:4))
}
\arguments{
\item{regi_type}{Type of registry.}
}
\value{
Character vector
}
\description{
Get the areacodes of the registry
}
