% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{std_lang}
\alias{std_lang}
\title{Standardize language input}
\usage{
std_lang(lang)
}
\arguments{
\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}
}
\value{
Standardized language code: "en", "cn", or "code".
}
\description{
Standardize language input
}
