% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{strat_vars}
\alias{strat_vars}
\title{strat_vars}
\arguments{
\item{...}{One or more variables used for stratification. For example, you
can stratify by \code{sex}, \code{year}, \code{cancer}, or just by \code{year}. If sex
is not passed as a parameter, the output will be the result for the
combined gender.}
}
\description{
strat_vars
}
\keyword{internal}
