% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.canreg}
\alias{summary.canreg}
\alias{summary.canregs}
\alias{summary.check}
\title{Summary object of class 'canreg'}
\usage{
\method{summary}{canreg}(object, ...)

\method{summary}{canregs}(object, collapse = TRUE, ...)

\method{summary}{check}(object, ...)
}
\arguments{
\item{object}{Object data with class of 'canreg', 'canregs'}

\item{...}{Other filter expressions}

\item{collapse}{Collapse data or not.}
}
\value{
A data frame contains summary statistics of canreg data.
}
\description{
Summary object of class 'canreg'
}
