% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg.R
\docType{class}
\name{SWcases-class}
\alias{SWcases-class}
\title{Class of SWcases}
\value{
Class of SWcases
}
\description{
Class of SWcases
}
\section{Slots}{

\describe{
\item{\code{registr}}{character.}

\item{\code{id}}{character.}

\item{\code{name}}{character.}

\item{\code{sex}}{factor.}

\item{\code{age}}{numeric.}

\item{\code{agegrp}}{factor.}

\item{\code{birthda}}{Date.}

\item{\code{topo}}{character.}

\item{\code{morp}}{character.}

\item{\code{beha}}{character.}

\item{\code{icd10}}{character.}

\item{\code{inciden}}{Date.}

\item{\code{deathda}}{Date.}

\item{\code{areacode}}{character.}
}}

