% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_address}
\alias{tidy_address}
\title{Query address code}
\usage{
tidy_address(x, unique = TRUE)
}
\arguments{
\item{x}{A string vector that describe the address.}

\item{unique}{Logical value for multiple output or not.}
}
\description{
Query address code
}
