% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_age}
\alias{tidy_age}
\title{Tidy age description.}
\usage{
tidy_age(x, unit = "year")
}
\arguments{
\item{x}{Vector contains age description in Chinese.}

\item{unit}{Character, unit of values return, options are "year", "month",
or "day", default is "year".}
}
\value{
Numeric vector contains age.
}
\description{
Tidy age description.
}
\examples{
agedes <- c("50\u5c8110\u67083\u6708", "19\u5c815\u6708",
            "1\u5c8130\u6708", "3\u670820\u6708","30\u6708")
tidy_age(agedes, unit = "year")
tidy_age(agedes, unit = "month")
tidy_age(agedes, unit = "day")
}
