% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_areacode}
\alias{tidy_areacode}
\title{Get attributes affiliated with the areacode}
\usage{
tidy_areacode(x, lang = "cn")
}
\arguments{
\item{x}{Character vector of six digits areacode.}

\item{lang}{Character string indicate the language.}
}
\value{
List of atttributes of areacode.
}
\description{
Get attributes affiliated with the areacode
}
