% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_occu}
\alias{tidy_occu}
\title{Tidy occupation codes or description.}
\usage{
tidy_occu(x, lang = "cn")
}
\arguments{
\item{x}{A character vector contains description of occupation}

\item{lang}{Language, Options are "eng" for English, and "cn" for chinese.
Default is cn.}
}
\value{
A factor vector contains formatted occupation.
}
\description{
Tidy occupation codes or description.
}
