% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_sex}
\alias{tidy_sex}
\title{Tidy gender variable}
\usage{
tidy_sex(x, lang = "cn")
}
\arguments{
\item{x}{Vector contains gender information.}

\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}
}
\value{
A factor vector contains gender information.
}
\description{
Tidy gender variable
}
\examples{
gender <- c("男", "男性", "女", "女性", "女子", "男子")
tidy_sex(gender)
}
