% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_var}
\alias{tidy_var}
\title{Convert variable according to the Standard for dataset of CR in China}
\usage{
tidy_var(
  x,
  var_name = "occu",
  label_type = "full",
  lang = "code",
  as_factor = FALSE
)
}
\arguments{
\item{x}{Character values of variable to be converted.}

\item{var_name}{Character value indicate name of the variable.}

\item{label_type}{Type of the label used ("full" or "abbr").}

\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}

\item{as_factor}{Logical, indicate whether output value as factor.}
}
\value{
A formatted value.
}
\description{
Convert variable according to the Standard for dataset of CR in China
}
