% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canreg_report.R
\name{validate_canreg_report}
\alias{validate_canreg_report}
\title{Validate a \code{canreg_report} object}
\usage{
validate_canreg_report(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, invisibly, if valid.
}
\description{
Validate a \code{canreg_report} object
}
