% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{var_map_list}
\alias{var_map_list}
\title{Generate variable mapping list}
\usage{
var_map_list(var_name)
}
\arguments{
\item{var_name}{Character value indicating the name of the variable.}
}
\value{
A list where each code corresponds to a vector of possible values.
}
\description{
Generate variable mapping list
}
