% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_registry.R
\name{write_area_type}
\alias{write_area_type}
\title{Write area type to the user defined dict.}
\usage{
write_area_type(areacode, area_type)
}
\arguments{
\item{areacode}{Six-digit administrative division code.}

\item{area_type}{Area type 'Urban' or 'Rural' for the input Six-digit
administrative division code.}
}
\value{
Null
}
\description{
Write area type to the user defined dict.
}
