% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_registry.R
\name{write_registry}
\alias{write_registry}
\title{Write the registry definition to the user defined dict.}
\usage{
write_registry(areacode, registry)
}
\arguments{
\item{areacode}{Six-digit administrative division code.}

\item{registry}{The name of the registry corresponding to the six-digit
administrative division code.}
}
\value{
Null
}
\description{
Write the registry definition to the user defined dict.
}
