% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_deal.R
\name{add_labels}
\alias{add_labels}
\title{Add variable labels for data set.}
\usage{
add_labels(x, label_type = "full", lang = "zh")
}
\arguments{
\item{x}{A data frame or tibble to be labeled.}

\item{label_type}{Type of label for the cancer variable. Options are:
\itemize{
\item "full" - Full descriptive labels.
\item "abbr" - Shortened labels.
\item "icd10" - ICD-10 codes only.
}
The default is "full".}

\item{lang}{Language of the labels. Options are:
\itemize{
\item "zh" - Chinese.
\item "en" - English.
}
The default is "zh".}
}
\value{
A data frame or tibble with labeled variables.
}
\description{
This function labels variables such as \code{sex} and \code{cancer} in the dataset
based on the specified language and label type.
}
