% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate the actual age between two dates}
\usage{
calc_age(birth_date, onset_date)
}
\arguments{
\item{birth_date}{Birth date.}

\item{onset_date}{Event date.}
}
\value{
Vector of ages.
}
\description{
Calculate the actual age between two dates
}
\examples{
sdate <- as.Date("1960-01-01")
edate <- as.Date("1980-12-31")
bdate <- sample(seq(sdate, edate, by = "1 day"), 100, replace = TRUE)
event <- sample(seq(as.Date("2020-01-01"),
                    as.Date("2023-07-08"),
                    by = "1 day"),
                100, replace = TRUE)
ages <- calc_age(bdate, event)
ages
}
