% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canregtools-package.R
\docType{package}
\name{canregtools-package}
\alias{canregtools}
\alias{canregtools-package}
\title{canregtools: Analysis Data from PBCR (Population-Based Cancer Registration,PBCR)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package includes functions for data analysis, visualization, and automatic report generation in the field of Population-Based Cancer Registration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gigu003.github.io/canregtools/}
}

}
\author{
\strong{Maintainer}: Qiong Chen \email{chenqiong@hnccr.org.cn} (\href{https://orcid.org/0000-0003-2401-0046}{ORCID})

}
\keyword{internal}
