% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canregtools-package.R
\docType{package}
\name{canregtools-package}
\alias{canregtools}
\alias{canregtools-package}
\title{canregtools: Analyzing Population-Based Cancer Registry Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools designed to streamline data analysis, visualization, and reporting in Population-Based Cancer Registration (PBCR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gigu003.github.io/canregtools/}
  \item \url{https://canregtools.chenq.site}
  \item Report bugs at \url{https://github.com/gigu003/canregtools/issues}
}

}
\author{
\strong{Maintainer}: Qiong Chen \email{chenq08@126.com} (\href{https://orcid.org/0000-0003-2401-0046}{ORCID}) [copyright holder]

}
\keyword{internal}
