% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_childhood}
\alias{classify_childhood}
\title{Classify pediatric tumors according to the ICCC3 standards.}
\usage{
classify_childhood(topo, morp, beha, type = "sub", version = "v2005")
}
\arguments{
\item{topo}{Topography codes of ICDO3, it could be in the format of 'C15.6'
or C156'.}

\item{morp}{Morphology codes of ICDO3, it could be in the format of '8000' or
'M8140'.}

\item{beha}{Behaviour code of ICDO3.}

\item{type}{Type of Classification for the output, options are 'main' or
'sub', default is 'main'.}

\item{version}{The ICCC3 standards, "v2005" or "v2017".}
}
\value{
Recode of ICCC3 classification.
}
\description{
Classify pediatric tumors according to the ICCC3 standards.
}
\details{
This function classify the 'topo' and 'morp' parts of ICDO3 codes into
ICCC3 (International Classification of Childhood Cancer, Third edition).
}
