% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_icd10}
\alias{classify_icd10}
\title{Classify ICD10 Codes to Cancer Categories.}
\usage{
classify_icd10(
  x,
  cancer_type = "big",
  lang = "code",
  label_type = "abbr",
  as_factor = FALSE
)
}
\arguments{
\item{x}{The ICD10 codes of cancer part (C00-C98 and D00-D48) collected
by the Population-Based Cancer Registration (PBCR).}

\item{cancer_type}{A character string specifying the classification method
used to categorize ICD-10 codes. This determines how ICD-10 codes are
classified. Options include \code{"big"} (classify ICD-10 codes into 26
cancer categories), \code{"small"} (classify ICD-10 codes into 59 cancer
categories, more specific categories), \code{"system"} (classify ICD-10
codes into organ system), and \code{"gco"} (classify ICD-10 code into
cancer categories same as classification published by the Global
Cancer Observatory). This parameter is only available when the
input data is a vector of ICD-10 codes, or object with class of
\code{'canreg'} or \code{'canregs'}.}

\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}

\item{label_type}{Type of the label used ("full" or "abbr").}

\item{as_factor}{Logical, indicate whether output value as factor.}
}
\value{
Cancer code
}
\description{
Classify ICD10 codes into Cancer Categories according to the specified
category type and language.
}
\examples{
icd10 <- c("C15.2", "C33.4", "C80.9", "C26.2", "C16.3")
classify_icd10(icd10, cancer_type = "big")
}
