% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_canreg.R
\name{clean_canreg}
\alias{clean_canreg}
\alias{clean_canreg.FBcases}
\alias{clean_canreg.SWcases}
\alias{clean_canreg.population}
\title{Clean canreg data.}
\usage{
clean_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

\method{clean_canreg}{FBcases}(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

\method{clean_canreg}{SWcases}(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

\method{clean_canreg}{population}(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)
}
\arguments{
\item{x}{Data of class 'FBcases', 'SWcases' or 'population'.}

\item{cutage_method}{Method for grouping ages. Options are 'distance',
'interval', or 'quarter'. Default is 'distance'.}

\item{smooth_age}{Method for expanding age. Options are 'linear','natural',
or 'periodic'. Default is 'natural'.}

\item{breaks}{Cut points for age groups when 'cutage_method' is 'interval'.
Default is 'c(0, 15, 40, 65)'.}

\item{length}{Length of each age group when 'cutage_method' is 'distance'.
Default is 5.}

\item{maxage}{Start age of the last age groups. Default is 85.}

\item{sep_zero}{Logical value TRUE or FALSE, if the 0-year age group treated
as a separate group. Default is TRUE.}

\item{labels}{Labels for the age groups.}

\item{label_tail}{Tail of the labels.}

\item{cancer_type}{Specify the grouping method for tumors. Options are 'system',
'big', or 'small'. Default is 'big'.}
}
\value{
Class 'canreg'.
}
\description{
Clean canreg data.
}
