% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_canreg.R
\name{count_canreg}
\alias{count_canreg}
\title{Count the canreg data.}
\usage{
count_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  type = "big",
  lang = "cn"
)
}
\arguments{
\item{x}{Object data of class 'canreg'.}

\item{cutage_method}{Methods for Specifying Age Groups. Options are
"interval", "distance", or "quantile". Default is "distance".}

\item{smooth_age}{Methods for expanding age groups from 5 years to 1 year.}

\item{breaks}{Specicify the break points classify age groups when
cutage_method is 'interval'. Default is 'c(0, 15, 40, 65)'.}

\item{length}{Specify the length of each age group when cutage_method is
'distance'. Default is 5.}

\item{maxage}{Specify the max age of age group when cutage_method is
'distance'. Default is 85.}

\item{sep_zero}{Logical value, TRUE or FALSE, specifying whether to treat age
0 as a separate group. Default is TRUE.}

\item{labels}{Labels for age groups. Default is NULL.}

\item{label_tail}{Tail label to be added to the labels. Default is NULL.}

\item{type}{Method for classify cancer sites.}

\item{lang}{Character value for specify the language used. options are 'cn'
for Chinese, or 'en' for English.}
}
\value{
An object of 'fbswicd'.
}
\description{
Count the canreg data.
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
fbsw <- count_canreg(data, cutage_method = "interval")
fbsw
}
