% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_canreg.R
\name{count_canreg}
\alias{count_canreg}
\alias{count_canreg.canregs}
\alias{count_canreg.canreg}
\title{Count canreg data.}
\usage{
count_canreg(
  x,
  cutage_method = "distance",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

\method{count_canreg}{canregs}(x, ...)

\method{count_canreg}{canreg}(
  x,
  cutage_method = "distance",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)
}
\arguments{
\item{x}{Object data of class 'canreg' or 'canregs'.}

\item{cutage_method}{Methods for Specifying Age Groups. Options are
"interval", "distance", or "quantile". Default is "distance".}

\item{breaks}{Specify the break points classify age groups when
cutage_method is 'interval'. Default is 'c(0, 15, 40, 65)'.}

\item{length}{Specify the length of each age group when cutage_method is
'distance'. Default is 5.}

\item{maxage}{Specify the max age of age group when cutage_method is
'distance'. Default is 85.}

\item{sep_zero}{Logical value, TRUE or FALSE, specifying whether to treat age
0 as a separate group. Default is TRUE.}

\item{labels}{Labels for age groups. Default is NULL.}

\item{label_tail}{Tail label to be added to the labels. Default is NULL.}

\item{cancer_type}{A character string specifying the classification method
used to categorize ICD-10 codes. This determines how ICD-10 codes are
classified. Options include \code{"big"} (classify ICD-10 codes into 26
cancer categories), \code{"small"} (classify ICD-10 codes into 59 cancer
categories, more specific categories), \code{"system"} (classify ICD-10
codes into organ system), and \code{"gco"} (classify ICD-10 code into
cancer categories same as classification published by the Global
Cancer Observatory). This parameter is only available when the
input data is a vector of ICD-10 codes, or object with class of
\code{'canreg'} or \code{'canregs'}.}

\item{...}{Parameters.}
}
\value{
data of class fbswicd.
}
\description{
Count canreg data.
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
fbsw <- count_canreg(data, cutage_method = "interval")
fbsw
}
