% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_canreg.R
\name{count_canreg}
\alias{count_canreg}
\alias{count_canreg.canregs}
\alias{count_canreg.canreg}
\title{Count canreg data.}
\usage{
count_canreg(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

\method{count_canreg}{canregs}(x, ...)

\method{count_canreg}{canreg}(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)
}
\arguments{
\item{x}{Object data of class 'canreg' or 'canregs'.}

\item{age_breaks}{Specify the break points classify age groups. Default is
\code{c(0, 1, seq(0, 85, 5))}.}

\item{label_tail}{Tail label to be added to the labels. Default is NULL.}

\item{cancer_type}{A character string specifying the classification method
used to categorize ICD-10 codes. This determines how ICD-10 codes are
classified. Options include \code{"big"} (classify ICD-10 codes into 26
cancer categories), \code{"small"} (classify ICD-10 codes into 59 cancer
categories, more specific categories), \code{"system"} (classify ICD-10
codes into organ system), and \code{"gco"} (classify ICD-10 code into
cancer categories same as classification published by the Global
Cancer Observatory). This parameter is only available when the
input data is a vector of ICD-10 codes, or object with class of
\code{'canreg'} or \code{'canregs'}.}

\item{...}{Parameters.}
}
\value{
data of class fbswicd.
}
\description{
Count canreg data.
}
\examples{
library(canregtools)
data("canregs")
fbsw <- count_canreg(canregs)
fbsw
}
