% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_merge.R
\name{cr_merge}
\alias{cr_merge}
\alias{cr_merge.canregs}
\alias{cr_merge.asrs}
\alias{cr_merge.qualities}
\title{Merge elements from objects of class canregs, fbswicds, or asrs}
\usage{
cr_merge(data, nested = NULL)

\method{cr_merge}{canregs}(data)

\method{cr_merge}{asrs}(data)

\method{cr_merge}{qualities}(data)

cr_merge(data, nested = NULL)
}
\arguments{
\item{data}{list need to be merged.}

\item{nested}{Nested list.}
}
\value{
An object with merged elements.

object
}
\description{
Merge elements from object of class 'canregs', 'fbswicds', or 'asrs' based
on provided index or conditions. The class of the returned object will be the
same with the input object.
}
