% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_age_rate}
\alias{create_age_rate}
\title{Create age specific rate.}
\usage{
create_age_rate(
  data,
  ...,
  event = fbs,
  rks = rks,
  agegrp = agegrp,
  type = "long",
  mp = 1e+05,
  decimal = 2
)
}
\arguments{
\item{data}{Data frame of class 'fbswicd'.}

\item{...}{Variable name used for stratification.}

\item{event}{Event used for age specific rate.}

\item{rks}{Risk population.}

\item{agegrp}{Variable name indicate age groups.}

\item{type}{Data frame format long or wide.}

\item{mp}{Correction factor.}

\item{decimal}{Decimal places for rounding.}
}
\value{
Return a data frame of age specific rate.
}
\description{
Create age specific rate.
}
