% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_rate.R
\name{create_age_rate}
\alias{create_age_rate}
\alias{create_age_rate.canreg}
\alias{create_age_rate.fbswicds}
\alias{create_age_rate.fbswicd}
\title{Create age specific rate}
\usage{
create_age_rate(
  x,
  ...,
  event = fbs,
  cancer_type = "big",
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)

\method{create_age_rate}{canreg}(x, ..., cancer_type = "big")

\method{create_age_rate}{fbswicds}(
  x,
  ...,
  event = fbs,
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)

\method{create_age_rate}{fbswicd}(
  x,
  ...,
  event = fbs,
  format = "long",
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE
)
}
\arguments{
\item{x}{The input data, object with class of \code{'fbswicd'}, \code{'fbswicds'},
\code{'canreg'}, or \code{'canregs'}.}

\item{...}{One or more variables used for stratification. For example, you
can stratify by \code{sex}, \code{year}, \code{cancer}, or just by \code{year}. If sex
is not passed as a parameter, the output will be the result for the
combined gender.}

\item{event}{A variable used to specify the type of calculation, options are
fbs or sws, fbs for cancer incidence, and sws for cancer mortality.}

\item{cancer_type}{A character string specifying the classification method
used to categorize ICD-10 codes. This determines how ICD-10 codes are
classified. Options include \code{"big"} (classify ICD-10 codes into 26
cancer categories), \code{"small"} (classify ICD-10 codes into 59 cancer
categories, more specific categories), \code{"system"} (classify ICD-10
codes into organ system), and \code{"gco"} (classify ICD-10 code into
cancer categories same as classification published by the Global
Cancer Observatory). This parameter is only available when the
input data is a vector of ICD-10 codes, or object with class of
\code{'canreg'} or \code{'canregs'}.}

\item{format}{Format of the output data frame, either "long" or "wide".}

\item{mp}{A constant to multiply rates by (e.g. mp=1000 for rates per 1000).}

\item{decimal}{This parameter specifies the number of decimal places to
round the results. The default is 2, which means rates will
be rounded to two decimal places.}

\item{show_pop}{Logical value whether output population or not.}
}
\value{
A data frame of age-specific rates.
}
\description{
Creates a data frame of age-specific rates based on the class of data
(\code{fbswicd} or \code{canreg}).
}
