% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_asr.R
\name{create_asr}
\alias{create_asr}
\alias{create_asr.canregs}
\alias{create_asr.canreg}
\alias{create_asr.fbswicds}
\alias{create_asr.fbswicd}
\title{Calculate age-standardized rate.}
\usage{
create_asr(
  x,
  ...,
  event = fbs,
  std = c("cn2000", "wld85"),
  mp = 1e+05,
  decimal = 2,
  show_var = FALSE,
  show_ci = FALSE
)

\method{create_asr}{canregs}(x, ..., cancer_type = "big")

\method{create_asr}{canreg}(x, ..., cancer_type = "big")

\method{create_asr}{fbswicds}(
  x,
  ...,
  event = fbs,
  std = c("cn2000", "wld85"),
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE,
  show_var = FALSE,
  show_ci = FALSE
)

\method{create_asr}{fbswicd}(
  x,
  ...,
  event = fbs,
  std = c("cn2000", "wld85"),
  mp = 1e+05,
  decimal = 6,
  show_pop = FALSE,
  show_var = FALSE,
  show_ci = FALSE
)
}
\arguments{
\item{x}{data with class of 'fbswicd' or 'canreg'.}

\item{...}{Variable names used to stratification.}

\item{event}{A variable within the input data that corresponds to the cases
counts.}

\item{std}{Specify the standard population structure in the 'std_pop'
data frame used for calculating standardized rates. When
calculating standardized rates for multiple standard populations,
specify std = c(segi, china).}

\item{mp}{A constant to multiply rates by (e.g. mp=1000 for rates per 1000).}

\item{decimal}{This parameter specifies the number of decimal places to
round the results. The default is 2, which means rates will
be rounded to two decimal places.}

\item{show_var}{Logical value whether output variance or not.}

\item{show_ci}{Logical value whether output confidence(lower or upper
bound) or not.}

\item{cancer_type}{Classification type of cancer site, options are "big", "small"
or "system", default is "big".}

\item{show_pop}{Logical value whether output population or not.}
}
\value{
A data frame or tibble contains the age standard rates and CIs.
}
\description{
Calculate age-standardized rate.
}
