% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map.}
\usage{
create_map(
  x,
  values = NULL,
  breaks = NULL,
  bg_level = "province",
  bg_bound = "county",
  bg_cols = c("lightgray", "white"),
  h_type = "",
  h_cols = c("white", "darkred"),
  label = FALSE,
  label_size = 2,
  legend = FALSE,
  legend_title = "Incidence"
)
}
\arguments{
\item{x}{Areacodes.}

\item{values}{values.}

\item{breaks}{Cutoff values.}

\item{bg_level}{Province or city}

\item{bg_bound}{Show bound of background map.}

\item{bg_cols}{background colors.}

\item{h_type}{map type.}

\item{h_cols}{highlight colors.}

\item{label}{Logical value.}

\item{label_size}{size of label}

\item{legend}{Logical value.}

\item{legend_title}{Title of legend.}
}
\value{
A map
}
\description{
Create map.
}
