% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create map.}
\usage{
create_map(
  x,
  value = NULL,
  breaks = NULL,
  class = "province",
  type = "ggplot2",
  bg_cols = c("lightgray", "white"),
  highlight_cols = c("white", "red"),
  label = TRUE,
  label_size = 2,
  legend = TRUE,
  legend_title = "Incidence"
)
}
\arguments{
\item{x}{Areacodes.}

\item{value}{Values.}

\item{breaks}{Cutoff values.}

\item{class}{Province or city}

\item{type}{ggplot2 or leaflet}

\item{bg_cols}{background colors.}

\item{highlight_cols}{highlight colors.}

\item{label}{Logical value.}

\item{label_size}{size of label}

\item{legend}{Logical value.}

\item{legend_title}{Title of legend.}
}
\value{
A map
}
\description{
Create map.
}
