% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_quality}
\alias{create_quality}
\title{Create quality index from object of class 'fbswicd'.}
\usage{
create_quality(data, ..., decimal = 2)
}
\arguments{
\item{data}{An object of class 'fbswicd'.}

\item{...}{Variables used for stratification.}

\item{decimal}{Decimal digits.}
}
\value{
Data frame of class quality.
}
\description{
Create quality index from object of class 'fbswicd'.
}
\examples{
library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
fbsw <- count_canreg(data, cutage_method = "interval")
quality <- create_quality(fbsw, year, sex, icd_cat)
}
