% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_barchart.R
\name{draw_barchart}
\alias{draw_barchart}
\title{Draw group bar chart.}
\usage{
draw_barchart(
  data,
  plot_var = cr,
  cate_var = cancer,
  group_var = year,
  side_var = sex,
  side_label = NULL,
  bar_side = 1,
  topn = 10,
  grid = c(1, 1),
  ...
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{plot_var}{Var to be plotted.}

\item{cate_var}{Category variable name.}

\item{group_var}{Group variable name.}

\item{side_var}{Panel variable name.}

\item{side_label}{Labels for each side.}

\item{bar_side}{Orientation of the bar chart: 1 for left and 2 for right.}

\item{topn}{Top n 'plot_var' in each 'group_var' and 'side_var'.}

\item{grid}{The number of rows and columns in the plot area.}

\item{...}{Other options in draw_bar.}
}
\value{
Bar chart
}
\description{
Draw group bar chart.
}
\examples{
data <- load_canreg()
rate <- create_asr(data, year, sex, cancer, event = fbs)
draw_barchart(rate)
}
