% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_barchart.R
\name{draw_barchart}
\alias{draw_barchart}
\title{Draw group bar chart.}
\usage{
draw_barchart(
  data,
  plot_var = cr,
  cate_var = icd_cat,
  group_var = year,
  side_var = sex,
  topn = 10,
  bar_side = 1,
  nrow = 2,
  ...
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{plot_var}{Var to be plotted.}

\item{cate_var}{Category variable name.}

\item{group_var}{Group variable name.}

\item{side_var}{Panel variable name.}

\item{topn}{Top n 'plot_var' in each 'group_var' and 'side_var'.}

\item{bar_side}{Bar oriention, 1 for left, 2 for right.}

\item{nrow}{The number of subplots included in each row.}

\item{...}{Other options in draw_bar.}
}
\value{
Bar chart
}
\description{
Draw group bar chart.
}
\examples{
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)
rate <- create_asr(data, year, sex, icd_cat, event = fbs)
draw_barchart(rate)
}
