% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{lt}
\alias{lt}
\title{Calculate life table from mortality rates.}
\usage{
lt(mx, sage = 0, agegrp = 5, sex = "male")
}
\arguments{
\item{mx}{Death rate at age x.}

\item{sage}{Start age.}

\item{agegrp}{Age groups width.}

\item{sex}{Gender for the life table, it could be "male","female" or "total".}
}
\value{
A data frame containing the following values.
\item{age}{Start age in each age group.}
\item{mx}{Death rate at age x.}
\item{qx}{The probability that an individual of exact age x will die before exact age x+1.}
\item{lx}{Number of survivors to exact age x. The radix is 1.}
\item{dx}{The number of deaths between exact age x and x+1.}
\item{Lx}{Number of years lived between age x and x+1.}
\item{Tx}{Number of years lived after exact age x.}
\item{ex}{Remaining life expectancy at exact age x.}
}
\description{
Calculate life table from mortality rates.
}
\examples{
px <- c(
  20005, 86920, 102502, 151494, 182932,
  203107, 240289, 247076, 199665, 163820,
  145382, 86789, 69368, 51207, 39112, 20509,
  12301, 6586, 1909
)
dx <- c(
  156, 58, 47, 49, 48, 68, 120, 162, 160, 294,
  417, 522, 546, 628, 891, 831, 926, 731, 269
)
mx <- dx / px
lt(mx, sage = 0, agegrp = 5, sex = "total")
}
