% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_canreg.R
\name{read_canreg}
\alias{read_canreg}
\title{Read cancer registration data in the format reported to the National Cancer
Center.}
\usage{
read_canreg(x, pop_type = "long")
}
\arguments{
\item{x}{Address of cancer registration data, it could be a specific excel
file, or a folder contains a series of Excel files.}

\item{pop_type}{Format of population sheet, options area 'long' or 'wide',
default is 'long'.}
}
\value{
An object of canreg class or a list of objects of canreg class.
}
\description{
Read cancer registration data in the format reported to the National Cancer
Center.
}
