% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reframe_fbswicd.R
\name{reframe_fbswicd}
\alias{reframe_fbswicd}
\title{Reframe the fbswicds.}
\usage{
reframe_fbswicd(x, ...)
}
\arguments{
\item{x}{Object of class fbswicds.}

\item{...}{Stratification variables used to reframe.}
}
\value{
Object of class fbswicd.
}
\description{
Reframe the fbswicds.
}
