% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_registry.R
\name{show_registry}
\alias{show_registry}
\title{Title}
\usage{
show_registry(registry_type = "all", source = "system")
}
\arguments{
\item{registry_type}{registry type}

\item{source}{\code{character} source of areacode}
}
\value{
Character vector
}
\description{
Title
}
