% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.canreg}
\alias{summary.canreg}
\alias{summary.canregs}
\title{Summary object of class 'canreg'}
\usage{
\method{summary}{canreg}(object, collapse = FALSE, ...)

\method{summary}{canregs}(object, collapse = TRUE, ...)
}
\arguments{
\item{object}{Object data with class of 'canreg', 'canregs'}

\item{collapse}{Collapse data or not.}

\item{...}{Other filter expressions}
}
\value{
A data frame contains summary statistics of canreg data.
}
\description{
Summary object of class 'canreg'
}
\examples{
data("canregs")
data <- canregs[[1]]
summary(data)


summary(canregs)

}
