% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_address}
\alias{tidy_address}
\title{Query address code}
\usage{
tidy_address(x, api_key)
}
\arguments{
\item{x}{A string vector that describe the address.}

\item{api_key}{api_key for service provider.}
}
\value{
A data frame contains formatted address information.
}
\description{
Query address code
}
