% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_var}
\alias{tidy_var}
\title{Reformat variable values for Cancer Registration in China}
\usage{
tidy_var(
  x,
  var_name = "occu",
  label_type = "full",
  lang = "code",
  as_factor = FALSE
)
}
\arguments{
\item{x}{A character vector containing raw values of a variable used in
cancer registry data.}

\item{var_name}{A character string indicating the name of the variable to
reformat (e.g., \code{"occu"} for occupation). Must be one of the variable
names defined in \code{tidy_var_maps}.}

\item{label_type}{Type of the label used ("full" or "abbr").}

\item{lang}{Character, specify the output language, options are 'cn',
or 'en', default is 'cn'.}

\item{as_factor}{Logical, indicate whether output value as factor.}
}
\value{
A character or factor vector of reformatted values. The output
depends on the settings for \code{label_type}, \code{lang}, and \code{as_factor}:
\itemize{
\item If \code{as_factor = FALSE}, returns a character vector.
\item If \code{as_factor = TRUE}, returns a factor with sorted unique levels.
\item The labels used depend on \code{lang} (\code{"cn"}, \code{"en"}, \code{"code"}, or \code{"icd10"})
and \code{label_type} (\code{"full"} or \code{"abbr"}).
}
}
\description{
Standardizes and labels values of a specified variable according to the
national cancer registration standard of China: T/CHIA 18-2021.
}
\details{
\code{tidy_var()} converts raw character inputs into standardized labels, codes,
or abbreviations based on reference mappings defined for each variable (e.g.,
occupation, basis of diagnosis, etc.). It supports both Chinese and English
outputs and can return values as factors with labeled levels.
}
\examples{
occu <- c("11", "13", "17", "21", "24", "27", "31", "37", "51", "80", "90")
tidy_var(occu, var_name = "occu", lang = "cn")
tidy_var(occu, var_name = "occu", lang = "en")
tidy_var(occu, var_name = "occu", lang = "cn", label_type = "abbr")
tidy_var(occu, var_name = "occu", lang = "en", label_type = "abbr")

}
