% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_registry.R
\name{write_areacode}
\alias{write_areacode}
\title{Write Six-digit administrative division code to the user defined dict.}
\usage{
write_areacode(x = NULL)
}
\arguments{
\item{x}{A data frame contains \code{areacode}, \code{cname}, \code{ename}, \code{abbr_cn}
\code{abbr_en}}
}
\value{
Null
}
\description{
Write Six-digit administrative division code to the user defined dict.
}
