% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_registry.R
\name{write_areacode}
\alias{write_areacode}
\title{Write Six-digit administrative division code to the user defined dict.}
\usage{
write_areacode(areacode, name)
}
\arguments{
\item{areacode}{Six-digit administrative division code.}

\item{name}{Label for the Six-digit administrative division code.}
}
\value{
Null
}
\description{
Write Six-digit administrative division code to the user defined dict.
}
